split forge's install mod
